+++ lib/src/models/company.dart
@@ -1,4 +1,4 @@
-import 'package:tmdb_dart/tmdb_dart.dart';
+import '../settings/asset-resolver.dart';
 
 class Company {
   int id;