+++ lib/src/models/keyword.dart
@@ -1,0 +1,13 @@
+
+class Keyword {
+  int id;
+  String name;
+
+  Keyword.fromJson(Map<String, dynamic> map) {
+    id = map["id"];
+    name = map["name"];
+  }
+
+  static List<Keyword> listFromJson(List<dynamic> list) =>
+      list.map((x) => new Keyword.fromJson(x)).toList(growable: false);
+}