+++ lib/src/services/tmdb-service.dart
@@ -1,7 +1,14 @@
 import 'dart:async';
 import 'dart:convert';
 import 'package:http/http.dart';
-import 'package:tmdb_dart/tmdb_dart.dart';
+import '../models/movie-base.dart';
+import '../models/movie.dart';
+import '../models/paged-result.dart';
+import '../settings/asset-resolver.dart';
+import '../settings/configuration.dart';
+import '../settings/movie-append-settings.dart';
+import '../settings/movie-search-settings.dart';
+import '../settings/quality-settings.dart';
 
 part 'configuration-service.dart';
 part 'movie-service.dart';