+++ CHANGELOG.md
@@ -16,3 +16,20 @@
 * Partial support for response appending of getMovie(), currently only Images are appended 
 * Renamed Resolution to AssetQuality
 * Extracted Asset Path resolving methods from Configuration class to AssetResolver
+
+## 0.0.4
+
+* Changed http package version 
+
+## 0.0.5
+
+* Upgraded version of the 'http' package
+* Upgraded sdk version to stable (was 'dev' previosly)
+* Enhanced getMovie() method to support new appending options
+    1. Alternative Titles
+    2. Credits
+    3. External Ids
+    4. Keywords
+    5. Videos
+    6. Recommendations
+    7. Similar Movies